Repository setup and history cleanup
We have migrated the project to a monorepo structure.
The repository, which initially contained only hardware drawings,
now includes all project components, organized by modules.

As part of this change, we also cleaned up the Git history.
While the original history was not critical early on, it became increasingly
cluttered over time.

This PR represents a breaking, irreversible change and should be considered a
point of no return.